+++ backend/src/main/java/payments/entity/PaymentEntity.java
@@ -6,7 +6,6 @@ import lombok.EqualsAndHashCode;
 import lombok.ToString;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -28,7 +27,7 @@ public class PaymentEntity {
 
     @JsonProperty
     @Column(precision = 12, scale = 2)
-    private BigDecimal sum;
+    private Long sum;
 
     @JsonProperty
     @ManyToOne
